feat(tasks): add trigger configuration and TypeScript support
