+++ .github/workflows/release-trigger-prod.yml
@@ -29,7 +29,7 @@ jobs:
         run: pnpm install --frozen-lockfile
 
       - name: 🚀 Deploy Trigger.dev
-        working-directory: packages/agent
+        working-directory: packages/tasks
         env:
           TRIGGER_ACCESS_TOKEN: ${{ secrets.TRIGGER_ACCESS_TOKEN }}
         run: npx trigger.dev@4.1.2 deploy
