+++ apps/api/src/trigger/dto/directory-command.dto.ts
@@ -1,6 +1,6 @@
 import { Type } from 'class-transformer';
 import { IsEnum, IsObject, IsOptional } from 'class-validator';
-import { DirectoryCommandAction } from '@packages/agent/trigger';
+import { DirectoryCommandAction } from '@packages/agent/tasks';
 
 export class DirectoryCommandDto {
     @IsEnum(DirectoryCommandAction)
