+++ packages/agent/src/tasks/directory-command.types.ts
