+++ packages/agent/src/tasks/index.ts
@@ -1,1 +1,3 @@
-export * from '@trigger.dev/sdk';
+export * from './directory-command.types';
+export * from './directory-generation.types';
+export * from './directory-generation-dispatcher';
