--- packages/agent/src/tasks/trigger/index.ts
@@ -1,1 +1,0 @@
-export * from './directory-generation.task';
