--- packages/agent/src/trigger/index.ts
@@ -1,9 +1,0 @@
-export * from './trigger.module';
-export * from './trigger.service';
-export * from './trigger-worker.module';
-export * from './trigger-internal-api.client';
-export * from './trigger-generation.orchestrator';
-export * from './trigger-ai.module';
-export * from './trigger-items-generator.module';
-export * from './remote-directory-operations.service';
-export * from './directory-command.types';
