--- packages/agent/src/trigger/trigger-ai.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AiService } from '@src/ai/ai.service';
-
-@Module({
-    providers: [AiService],
-    exports: [AiService],
-})
-export class TriggerAiModule {}
