--- packages/agent/src/trigger/trigger.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TriggerService } from './trigger.service';
-
-@Module({
-    providers: [TriggerService],
-    exports: [TriggerService],
-})
-export class TriggerModule {}
