+++ packages/tasks/.gitignore
@@ -1,0 +1,3 @@
+tsconfig.tsbuildinfo
+# Trigger.dev
+.trigger
