+++ packages/tasks/src/index.ts
@@ -1,0 +1,1 @@
+export * from './trigger/trigger.module';
