+++ packages/tasks/src/trigger/cache/internal-api.adapter.ts
