+++ packages/tasks/src/trigger/trigger-ai.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { AiService } from '@packages/agent/ai';
+
+@Module({
+	providers: [AiService],
+	exports: [AiService]
+})
+export class TriggerAiModule {}
