+++ packages/tasks/src/trigger/trigger-internal.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { TriggerInternalApiClient } from './trigger-internal-api.client';
+
+@Module({
+	providers: [TriggerInternalApiClient],
+	exports: [TriggerInternalApiClient]
+})
+export class TriggerInternalModule {}
