+++ packages/tasks/src/trigger/trigger.module.ts
@@ -1,0 +1,16 @@
+import { Global, Module } from '@nestjs/common';
+import { TriggerService } from './trigger.service';
+import { DIRECTORY_GENERATION_DISPATCHER } from '@packages/agent/tasks';
+
+@Global()
+@Module({
+	providers: [
+		TriggerService,
+		{
+			provide: DIRECTORY_GENERATION_DISPATCHER,
+			useExisting: TriggerService
+		}
+	],
+	exports: [TriggerService, DIRECTORY_GENERATION_DISPATCHER]
+})
+export class TriggerModule {}
